SubstratumUI: Rebase Root commands to Kernel Auditor's version
Thanks to @USA-RedDragon for the files, and thanks to Kernel Auditor’s
open source work for Substratum’s root commands to be rebased.

Why: This creates one shell at the start of the app and will run it
under that shell instead of creating a new superuser shell every time a
new command needs to be run.