References: Move theme filter to References and change for 170
This marks the first specified date where themers must change their
metadata based on an announcement.

This will happen more frequently to keep themers up to date with the
latest Substratum source.

All strings are changeable from References.java which controls 16 other
files and it’s variables. If there is a need to change it in the
future, change it here.

NOTE TO DEVS: Never expose this to an array file as themers can
possibly change it - hardcode it using the References.java file.