AdvancedManagerFragment: Reintroduce uninstalls
This commit re-adds the functionality of uninstalling overlays directly
from within the Advanced Manager Fragment.

Themers: It is HIGHLY recommended that you inform your users to only
use this after they have DISABLED their overlays, or else it would
cause a huge issue with FCs. Although it is to be expected to have a
lot of FCs and we will look into this further later.