AntiPiracy: Marry the IsThisAMasquerade and AreYouSure libraries
Introduce another method of checking whether the classes are all
included. If IsThisAMasquerade fails, the AreYouSure library would
recheck the classes to ensure anti piracy measurements is kept.

If at any point the pirates reached the realization that there are two
filters, divorce the libraries and remove AreYouSure from the picture.