Substratum: Pre-compile 80% of all dependencies for NDK
This commit adds pre-compiled versions of the dependencies for
Substratum to function by building through NDK, but there are a few
more left that aren’t behaving properly when they are precompiled.

By precompiling many of these dependencies, we eliminate all the Gradle
Task errors as well as speed up overall time.