SubstratumBuilder: Allow for overlays to be built for Nougat
Android N introduces a new function to enforce the “hasCode” function
on the enforceCodePolicy() method. By default it is set to true, to
maximize compatibility with applications already on the Play Store.
However, overlay APKs do not have any code in them, hence they are
always treated as “hasCode=true” on Nougat. This commit cleans up that
functionality and forces it to say “hasCode=false” so the system will
let it install.