Notifications: Brand new UI to interact with Resource Initializer
Introduce a brand new way for the user to interact with the automated
resource cache initializer. The user can now cancel and expand the
themes at a later date.

A notification will pop up when a theme gets installed, and it will
show live progress of the expansion. Once complete, it dismisses itself
and shows the notification saying it was successful. There is no
failure notification for this.

This also allows the user to see when they clean the cache of the app.
If by any chance the user’s Substratum application terminates while the
notification is expanding (held on an X% for too long), the Cancel
button will automatically reset the state of “is_updating”, as to
increase the chances to prevent the “A theme is upgrading in the
background” toast.