OverlaysList: fail_count must be less than overlays building
This commit fixes an issue where when an two overlays (one works, one
is broken) are built, it should build and enable/update as usual. But
if we build one broken overlay, it proceeds to request masquerade to
enable it (with the progress bar). This is unintended behaviour and has
been fixed.