InformationTabsAdapter: Rewrite all the logic and simplify
This adapter took many code reviews and many people’s precious time but
finally after an epiphany I found one of the better ways of organizing
this to consolidate the whole logic into one proper function that
processes each tab dynamically.

The ArrayList will now be the temporary cache of what is “remaining” in
terms of what tabs have been instantiated prior to relaunching another
tab. If a tab is spawned, remove itself from the ArrayList so the
remaining tabs would not have a chance to respawn the same fragment
again.