SubstratumRebase: Consolidate Root commands into References
This rebase commits moves all of the root commands in Substratum into
one centralized area for further progress in the upcoming commits.

We should consolidate the ThreadRunner class but this is awaiting the
Masquerade overhaul to handle the OverlayManagerService natively.

Should one find themselves stumbling across this commit, this commit
allows for References.java to control most if not all root calls on the
app, and thus when the team implements the no-root system it would be
easy to transition over in a snap.