ThemeUninstallDetector: When themes uninstall, wipe the cache!
It is common that users start uninstalling themes from the Play Store,
or from Settings. Doing it this way, the cache isn’t wiped.

This commit creates a BroadcastReceiver to listen to theme uninstalls
and when it matches a recognized theme, purge the SubstratumBuilder
cache, ultimately freeing the space on the user’s device.