IconStudio: Initial contribution to the icon pack framework
- Properly buffers all the icon packs in an AsyncTask on a GridLayoutManager
RecyclerView layout, with 4 being the max amount of objects on the X scale.

- The FAB shows up in each of the preview activities

- The selector screen gets buffered within milliseconds on the packages
that use a launcher based metadata (Nova Launcher filter).

- Implemented function to quickly grab the exact locations within each
target APK (APK to receive an icon pack overlay) for its launcher icons,
e.g. if an app used /res/drawable-nodpi-v13/launcher_magnificent.png as
the launcher icon, Substratum now fully reads all of this and implements
it in a documented HashMap function for further calls and API hooks later
down the road.

- Replicate a new SubstratumBuilder to allow for icon packs to be compiled
specifically with respect to new metadata tags and package naming convention.

- Introduce the compilation system that compiles full packs at once

- Introduce a new blur dialog specifically for icon packs

- All downloaded icon packs will be scrutinized by Substratum. If there is a
rogue icon inside the icon pack, it will announce for the Icon Pack themer
where the issue is, for example, if an icon pack is missing a drawable resource
then it will complain under ResourceTypes that there is no package identifier
when getting value for a 0x000000000 resource number.

- The Icon Pack Studio framework will dynamically overlay all resolutions
when only one icon is applied, for example, icon packs usually have one
resolution for icons, 192x192. However, when overlaying to system, we need to
ensure that the icons get overlaid on all DPI scales, so it will be automatically
resized to fit each resolution respectively.

- The Icon Pack Studio will not overlay system applications, such as SystemUI,
even though Icon Pack themers do not theme these icons on any launcher anyways,
it has been added to the safeguard that these icons (SystemUI) are dangerous
to overlay simply with a drawable PNG, as the icons are linked with deep-framework
changes.

- The Icon Pack Studio will refuse to overlay an Icon Pack dashboard application,
as this seems redundant.

- The Icon Pack Studio will filter out only the visible applications with a valid
Java activity will be overlaid. As this is meant to theme "Launcher" icons, it makes
most sense that only visible icons be overlaid.

- The Icon Pack Studio will filter out all the activities found on the system and
dynamically decide whether which activity is "most-fit" for use by the Icon Pack
Studio using the Substratum Best Match algorith. This prevents the Settings icon
from turning into the Bluetooth icon, as many icon pack devs theme specific
activities occassionally.

- The Icon Pack Studio solely relies on OMS7, and masquerade 20 to communicate
specific commands to refresh the windows.

- Framework and Substratum overlays will trigger the update_mode flag, so we make sure
that it is set to be configured to the end onPostExecute, just before being sent
to masquerade for further instructions.

- Make sure Advanced Manager and OverlaysTab does not crash by using a filter.