SubstratumBuilder: Introduce target app referencing (AppCompat)
This commit allows for themers to call upon private resources defined by the public.xml file in the decompiled theme.

Theoretically, this means that themers can now use:

@*com.google.android.apps.plus:style/Theme.AppCompat
@*com.android.systemui:drawable/icon
@*com.android.systemui:string/data_usage_uninstalled_apps
@*com.android.systemui:dimen/fab_size
@*com.android:systemui:id/scrollView
@*com.android.systemui:plurals/status_bar_accessibility_recent_apps
@*com.android.systemui:integer/long_press_dock_anim_duration
@*com.android.systemui:layout/lb_header
@*com.android.systemui:attr/guidedActionDescriptionMinLines
@*com.android.systemui:interpolator/lockscreen_fingerprint_error_state_to_fp_animation_interpolator_0
@*com.android.systemui:array/color_mode_names
@*com.android.systemui:bool/config_sendPackageName

With the resolution of calling Theme.AppCompat as a parent by the self-referencing AOPT modifications, it would resolve issues where crashes occur when the style is "not complete".