Check for bootlooping ROMs with isIncompatibleFirmware
Instead of checking against a known set of values, consider every firmware after the November Security update to be guilty until proven innocent.
When (if...) Google merges a fix for this, we will need to come back and change the boolean result into a switch case (or equivalent) between the affected dates.