SubstratumRebase: We're now on steroids in terms of load speed!
Kill off non-6.0.0 theme templates in preparation for the month-long delay for themes. This completely kills load time when loading a different tab from the navigation drawer (e.g. Overlays, Fonts, Boot Animations, Sound Packs and Wallpapers)

This speeds things up by 50x original speeds (from 2 secs -> 0.04secs (40ms))

AOPT check now only happens on the Home Fragment to further optimize the fragment switching.

As we are only looking for specific packages under a specific intent filter, this no longer is an algorithm that is directly proportional to how many packages (input) are located on the device with a complexity of O(n), but rather a more refined algorithm that runs through a complexity of O(logn)