SubstratumUI: Interface directly to the OverlayManagerService
This commit takes the AOSP code of the OverlayManagerService, allowing
Android Studio and Substratum to interface directly to OMS, to query
and modify the database without root necessary.

This commit is a WIP and is not bounded to any functions of the app
just yet. Soon we will be tying the OverlayManagerService class to the
rest of the app in preparations for the no-root rebase.