MainActivity: Search query should not reload fragment on icon click
Add checks for when a user clicks on the search button and the fragment does an extra reload.

When the search button is clicked, the query immediately goes to "" and userInput (our global variable) has been changed, ultimately leading to one fragment change. Adding this commit will check whether the query is actual a real user query and thus reloading based on real input rather than an icon press.