MainActivity: Try a different search filter algorithm
When leaving using the back button, the search query was not cleared. If we initialize the userInput with a blank string, the same as query when the search button is cleared, there won't be a fragment reload on a blank string, and leaving back button will reset the user's search query.