SubstratumUI: When the cache breaks, let user wipe immediately
Before this patch, an ambiguous toast will say that it is upgrading in the background, but there is no notification.

Substratum will now check if there is a notification in SystemUI. If there is, then it is upgrading normally. If there isn't, then the cache wipe snackbar will pop up alerting the user to wipe the cache and reset flags, as they probably FCd Substratum while cache was upgrading.