Revert "MainActivity: Clear the /.substratum/ folder when starting the app"
There has been some cases where the first few overlays were updated, then MainActivity reloaded.

By the nature of window refreshes, it is like a stack where when an activity is refreshed of assets, each activity and fragment stack would be reloaded and simulated that they have been loaded consecutively in one flicker.

MainActivity was in charge of deleting overlay APKs and when it refreshed in InformationActivity, MainActivity reloaded it's onCreate method, thus loading the overlay folder remover.