FloatUI: Rewrite algorithm to show proper foreground tasks accurately
To replicate: While in FloatUI mode, swipe down the status bar/QS and if your finger slides to a different app's notification, the app detected would be different.

By implementing a new algorithm with usage delay, we are able to accurately monitor what is being shown as a foreground of the user's screen to allow FloatUI to properly identify the correct package name.