FileOperations: new rootless createNewFolder
Expand createNewFolder with rootless method, following the logic set by the
other methods found in FileOperations.
It requires the "mkdir" job commit in Masquerade.
It allows creating the "/data/system/theme" folder, that's why "MasqueradeJobService"
checks are added (so the root dialog doesn't appear in rootless mode).

a combined work with KreAch3R.