CacheCreator: Recalculate optimized SystemUI refresh modulo algorithm
We should try limit as many SystemUI leaks as possible. By limiting it down to notifying the notification once every 200 files extracted, it would not extract like crazy, when big themes like Swift or Spectrum with 17000 files+ gets extracted.

This will also allow the notification's button to be easily pressed if the user wants to cancel a huge operation.