InformationActivity: Revamp the favorites icon, introduce changelogging
Themers just have to add:

    <string-array name="ThemeChangelog">
        <item>Test1</item>
        <item>Test2</item>
        <item>Test3</item>
        <item>Test4</item>
        <item>Test5</item>
        <item>Test6</item>
        <item>Test7</item>
        <item>Test8</item>
        <item>Test9</item>
    </string-array>

Where each of the tests are an individual change the themer wants to inform the user. The String[] will automatically be parsed and thus be added with automatic newlines and bullet points.

Favorites icon will solidify or empty depending on whether the theme is favorited now.