ThemeManager: Introduce blacklisting overlays to safeguard the system
To pass Google Sony and LineageOS' CTS, we must have an implementation
that safeguards certain packages and services from being overlaid.

To add more packages, append to the end of the String[].

We must filter out the masquerade package so that rogue themers would
not intentionally try to disrupt a user's set up, ensuring that the
system's integrity is matched to team standards.