Change the showcase layout
As of right now this works as in if you plug in a background image, the layout and "look"
is complete. The problem is that the database needs an update in order to successfully
pull in the card background image. This is something that Nick and Nathan will decide
on how to do it.

My suggest is having themers contribute their own but if we can crawl the playstore listings
for the card background image then more power to you guys :-)

Ignore the second card, background image was hardcoded for preview
- http://i.imgur.com/7C7Tklq.png

[DO NOT MERGE WITHOUT UPDATING THE DATABASE]