BootAnimations: Rewrite algorithm to determine bootanimation previews
On many themes, there are bootanimations that are very small but contain a lot of frames. As the frames get larger, the amount of memory needing to play them gets higher, all the way till an OutOfMemoryException, ultimately telling the user that they ran out of memory.

On my Nexus 6P, the app's VM reached a whopping 770mb, then just outright rejected to play the bootanimation. What a waste! With this commit, the app will now slim dynamically check how many files there are in the boot animation cache, and determine using an algorithm whether the resampling size for the image has to be modified or not.

This patch went from a 770mb exploded VM to a steady 111mb maintained VM, with a properly playing bootanimation.