Check for ROM dirty flash and reset
Substratum is caching important checks about the ROM's configuration,
which are integral to its function.

A ROM flash can change, break or disable a lot of these checks.

Make Substratum aware of the build date of the ROM (universal
across all AOSP builds) and force it to clear its data if a
dirty flash is found, so it can run all the checks again on the
next run.

Conflicts:
	app/src/main/res/values/strings.xml