Showcase: #ThemerEquality
From now on, the showcase will not sort the list alphabetically by default (there's an icon on top to sort it alphabetically). By default, it will randomly jumble the ArrayList<ShowcaseItem> up so that every themer feels respected that their theme would have equal amount of chances to be shown at the top rather than adapting to the alphabetical nature.