Update non-Theme Ready alert dialog
Changed the alert dialog description from telling users to remove Theme Ready if they have it installed. This is an unnecessary demand for themers who do not support Theme Ready, because it will not affect any themes which do not require it to be used. Theme Ready supports the themers, they don't need to change anything in their themes to support it.