Add a basic indicator for whether a ROM is supported or not
Nick created a list of ROMs that were verified to have the necessary commits
to run Substratum properly with little to no jank. This 'mod' checks the
BUILD.PRODUCT prop against that list and determines which string to show in Settings.

This is done as simple as possible to not cause any inconvenience to the user. I do
advice it be kept like this because a dialog or anything beyond this will probably
annoy the user running an unsupported ROM.

If the user has a problem, the developers will ask for a screenshot of Settings
to see what the user is working with and this will show the developer at a glance
whether the ROM is on the list or not.

- I do not have access to all ROM build props so there's a chance there's wrong info
below. Please double check before merging