PackageModificationDetector: Completely revamp the detection module
We had many unnecessary BroadcastReceivers to check whether themes were
valid to be used. This commit removes all the unnecessary excess
checkers and updates to the latest algorithms to check for themes.

All packages are now dynamically checked using PackageManager's
EXTRA_REPLACING intent rather than us creating our own cache to see if
it is being updated or not