Themers, let's start getting on deck with Android O! Shall we?
Sorry for the delay, had exams to worry about.

Developers: We may want to look into an alternate approach when APIs are fully documented for Android O. The issue here was that the Stream was prematurely closed after performing a specific Process shell execution. 

The exact caught error is:

04-11 21:19:54.468 D/SubstratumBuilder: Overlay APK creation has completed!
04-11 21:19:54.469 W/System.err: java.io.IOException: Stream closed
04-11 21:19:54.469 W/System.err:     at java.lang.ProcessBuilder$NullOutputStream.write(ProcessBuilder.java:433)
04-11 21:19:54.469 W/System.err:     at java.io.OutputStream.write(OutputStream.java:116)

This commit was not necessary for Android N and below, however the highly undocumented API code for Android O may have changed a significant amount of things with regards to how things are being run.

Another look would be this log that complains after using the Process runtime execution:

04-11 20:47:47.623 E/Notification: Use of stream types is deprecated for operations other than volume control.
04-11 20:47:47.623 E/Notification: See the documentation of setSound() for what to use instead with android.media.AudioAttributes to qualify your playback use case