Wallpapers: Asciibetically sort the list of wallpapers excluding spaces
According to this table: https://unicode-table.com/en/

We are able to see that "~" has a lower priority than "-" which is used to distinguish between the name and the preview of an image. As we do not want a symbol that would be distinguished by the asciibetical organization and order the wallpapers incorrectly, we must use a lower priority symbol to replace spaces from the XML. In the app, it will automatically convert the "~" to a space accordingly.