If something is not supported is......unsupported
Let's not pretend like something is not supported. If a ROM is not on the supported list then is either not supported. 

Quit this 'is not known'. Yes there is a rare chance that a ROM may be supported and not on the list but at that point is just because the developer of said ROM is lazy or doesn't care for Substratum in the first place. Let's not encourage laziness or mislead Substratum users because we want spare feelings here.