FloatUI: Revamp the UI based on ManagerFragment
This also converts the UI to a recyclerView rather than a checked listview. Which is much better in terms of dynamism and memory efficiency.

I also renamed a few files because while trying to clean up the old floatui layouts, I kept seeing files I thought were part of it, but weren't.