References: Type Annotation to @NonNull for checkOMS()
By now we all know the fundamentals of Java8, however it should be noted that type annotations should not be used to enforce nullability. Which is why we have a secondary check to return true if the context is null, as we expect window refreshes to only be on OMS based ROMs.

However due to the nature of @NonNull type annotation, when our compiler detects a possible NPE inbound, it will warn us beforehand. While the code is running, it doesn't matter whether this type annotation is in place, a non null context can still be passed through, which is why our secondary check enforces it for us.