EncryptedAssets: Perform handshake with theme package to update overlays
To blindly update overlays, themes will have to willingly hand over a copy of their encrypted decryption key and IV spec decryption key directly to Substratum.

Usually we can do all of this all at once when we launch a theme in IA, but because in OverlayUpdater, we do not launch the theme at all, but instead, we just let it update overlays in the background, so these decryption keys must be willingly given to substratum through a secured receiver tunnel. 

Once the v10.0.0 theme template is compiled, it generates a random set of keys of which the themer nor lead developers will know, so if you do lose your project files, you will be SOL. However, if you keep your theme backed up in GitLab or BitBucket on a private repo, you will be safe.