SamsungIntegration: When overlays are installed, auto close processes
The broadcast intent PACKAGE_ADDED is protected by Samsung to require a certain permission. However this permission is not registered unless it is a platform/system application. Thus, it is not able to test on a non-Samsung device with regards to PACKAGE_ADDED for overlays. Use startActivityForResult() instead for Package Manager.