AOPTCheck: remove the memory overhead caused by the unnecessary check
since the value of integrityCheck will never be null, and since the check output is
redirected to stdout, it saves us a bufferedread.

Despite the memory overhead savings being minimal, the check in and of itself is essentially
whether the user is selecitng stable orcutting edge version. So lets remove the check.