Overlays: split out Overlays java
the Overlays class under tabs houses the majority of Substratums compile commands
and as is, is very crowded. This temporarily makes the method declaration public
to help spit out Overlays into seperate classes to help make it easier to read and
edit.

Eventually may consider making compile commands an interface