ProfileFragment: Fix code duplication
When the profile_selector Spinner is instantiated, an ArrayList
and ArrayAdapter are created for it to inflate values from. The
code that populates the ArrayList was duplicated from the RefreshSpinner
method. We create and assign the ArrayList and ArrayAdapter as blank
entities then call the RefreshSpinner method to populate the ArrayList,
allowing us to get rid of the duplicate code.