SamsungIntegration: Allow for users with root to do things silently
This will batch uninstall/install overlays silently without requesting the user to click on install for each overlay.

This also adds a service for Samsung devices to monitor package changes. Because PACKAGE_ADDED is not broadcasted on a Samsung device when an overlay is installed, we must do this manually in the background.