SubstratumCrash: When the app crashes randomly, catch it
This also checks if the crash was based on ResourceNotFoundException or InflateException, which we all know is what happens when an overlay targeting Substratum is out of date and the resource pointers are shifted in the wrong place. It will pulsate the Rescue Me! button to be eye catching for the user.