InformationActivity: Glide; do not use any cache
This should fix the sporadic events that plague the randomness of the hero image changing to another theme. Not only does this commit remove the static variables called from nested fragments in an activity, but it shifts them to bundles where all tabs can call the bundle thanks to the adapter.

The BootCompletedDetector will always ensure that the Glide cache will be cleared and memory will be wiped. No longer do any of the Glide loads cache onto the memory (except online to reduce multiple calls), but it ensures that it won't run off memory.

Finally, no longer do we call to initialize the hero image twice when launching themes (for the options menu icon color analysis) and thus it would allow bigger themes to load significantly faster.