SubstratumUX: If a theme was upgraded in the background, relaunch IA
So let's say Play Store automatically updated a theme while you have InformationActivity open, you don't want to see old resources such as the old hero image and the old overlay names and types. 

The activity will now automatically restart, as well as gracefully finishing if the user is compiling a theme at that time.

NOTE: If by some chance the user updates the theme during an overlay install session, and the theme updates in between, the subsequently compiled overlays will be up to date. The activity will restart after due to the gracefully finishing code.