SubstratumRebase: Full cleanup, ButterKnife, JavaDocs
This commit completely rebases the way Substratum is now formatted.

- Most crucial functions within activities/fragments with views are now
clearly documented with JavaDocs. This will set forth a better way of
documenting code for people outside of the team.

- Fix incorrect LG framework naming in one position by using our global
variable system in Resources.java.
Resources.java now houses a bunch of variables used to communicate
amongst Fragments and Activities, as well as common variables that can
now be directly referenced instead of written as a String. Meaning if
the format of a feature in our template is renamed, we can just change
one variable to save ourselves time.

- Integrate the ButterKnife dependency, of which was a long time need
from my end, but due to most contributors are AOSP/Platform devs, using
foreign externalized frameworks are rather alien to them. At this point
in time, we will start shifting Substratum to a more "reactive" future!

- Fixed a few miscellaneous incorrectly named directories in
ProfileFragment.

- Removed Force Stop Interfacer settings toggle, including all
subsequent properties

- Removed excessive use of "this." and "final" with the previous lint
commit pushes.

- Split up all the activity internal calls to a new enum named Internal.
This will house the variables shared between multiple fragments and
activities.

- Fix the main activity's search function after the DiffUtil commit, as
well as improve the accuracy of the ManagerFragment's search filters

- Fixed a bug where enabling an uncompiled overlay would show the
progress bar

- Fixed a bug where two lunchBars show subsequently when swapping states
in Overlays tab