FirebaseMessagingService: Deprecate notification pushing system
We no longer use this feature, and because there are so many people out
there too afraid to read what Firebase does for developers, might as
well remove this feature with a 10-ft pole.

Only the notification pushing system (from our server) would be
disabled, analytics and crash reports are still obtained with no
user-identifiable information.

Test: None

Signed-off-by: Nicholas Chum <nicholas@prjkt.io>