SettingsFragment: Improve how we restart Substratum after settings changes.
Whatever we have right now works, but we're always prone
to the usual fragment detachment bugs. Not fun. Killing
substratum entirely before we restart it via the intent
hooked into the AlarmManager gracefully works around this.

Restart code pulled from @deletescape's luncher ;)
TODO: Might want to work on the delays to make the
transition less jarring but this works for now.

Test: Toggle any settings that restarts substratum
Test: No fragment detached crashes observed.
Signed-off-by: Harsh Shandilya <harsh@prjkt.io>