OverlaysManager: Properly close compilingProgress
Turning this feature off was just temporary, but now themes will relaunch when you push a new update WHILE the theme is open (in recents, or while staring at your theme because we all know your SOT is 9hrs because all you do is ignore your real life friends to talk to pixels on your screen). This should also take care of any potential side problems with the relaunching mechanism.

Why do we need this? This is so that the entire theme will be reloaded just in case some people changed crucial elements of the theme, like type3s and changing the value of that from time to time. Also, hero images and all tabs will be reloaded.

Test: Open a test theme. Compile one overlay. Press home after compiled. Then, push an update test theme (no need to have any changes) and then your theme will relaunch including the full stack of activities of substratum, with a toast saying the theme was relaunched.

Test2: Open a test theme. Push an updated version of test theme (no need to have any changes) and the activity will close and relaunch with a toast saying the theme was relaunched.

Signed-off-by: Nicholas Chum <nicholas@prjkt.io>