ManagerAdapter: Fix theme entry formatting
Commit 702c2f67("ManagerAdapter: Fix inconsistent labels")
did fix what it advertised but also uncovered why the
"Theme: themeName" entry was not being stylised as
"<b>Theme</b>: themeName". Rather than doing the stylization
in the ManagerAdapter as we should have, we ended up doing
it in the ManagerItem _and_ calling toString() before we
returned the value from the getThemeName() method which
ended up with all stylization being lost.

Fix this by moving back the formatting into ManagerAdapter.

Fixes: 702c2f67("ManagerAdapter: Fix inconsistent labels")
Fixes: 7a61e1c9("ManagerAdapter: Fix incorrectly styled Theme Name entry")

Test: Visual.
Signed-off-by: Harsh Shandilya <harsh@prjkt.io>