OverlaysManager: If an overlay is about to be recompiled, disable first
To understand the core issue:

Compile an overlay with build and enable, then
with build and update. Said overlay will show as enabled
but would not actually be applied, making target package
appear unthemed.

This appears to be a callbacks handling error from OMS,
which has been rectified in https://substratum.review/#/458.
This patch aims to workaround this bug on stock ROMs.

Test: Running build and update on an already enabled overlay
will update the overlay correctly and the target package
will stay themed.

Signed-off-by: Harsh Shandilya <harsh@prjkt.io>