FragmentTransaction: Replace commit() with commitAllowingStateLoss()
This prevents people who intentionally try to break the app to break the
app, as it will allow the bundle to be reset

This commit also allows for one refresh after installing an overlay on
Samsung devices.

Test: On a Samsung device, select all overlays in a theme and click
Install Selected, then when the overlays ask to be installed, spam click
Cancel. The app will no longer crash.

Signed-off-by: Nicholas Chum <nicholas@prjkt.io>