OverlaysManager: Insert delay when installing snowflake overlays
When installing and enabling an overlay, we must take account for that
overlays cannot be enabled immediately after it was installed especially
when the OverlayManager service is activating the overlay. So, we add a
small delay before enabling snowflake overlays so that the user won't
have to re-enable them after.

Tested on OP5T with Edge theme - ~20 overlays

Test: Select all overlays (~20) in a theme and click Build and Enable.
At the end, all overlays should enable perfectly.

Signed-off-by: Nicholas Chum <nicholas@prjkt.io>