SamsungIntegration: Re-add overlay component permission for 7.1.1+
ATTENTION: Do not enable this for 7.0 devices as it will hide in
Settings and cause the overlay to go missing completely after a reboot.
We need confirmation that this works on 7.1.1 but afaik, Oreo it works
perfectly fine.

This results in overlays being active on first reboot and also stick on reboots.
Though, it breaks installation, requiring a follow-up commit.

Signed-off-by: Nicholas Chum <nicholas@prjkt.io>