OverlaysManager: Alert if user is installing overlay on stock March
This will show a dialog that tells the user to reboot when they see a
pink overlay for the first time. There is an option to hide the dialog
from ever showing again. This will not show on Substratum Service ROMs
and only on stock based ROMs such as Samsung Oreo and unsupported custom
ROMs like Lineage - with this alert, it provides the user with full
instructions to have overlays working after March patch.

Signed-off-by: Nicholas Chum <nicholas@prjkt.io>