AndromedaBinderService/Substratum: End binder services gracefully
At boot, if the binder fails, allow the user to execute it again by
launching the activity. This follows the natural human response of not
opening substratum BEFORE andromeda is connected, but if the human does,
then it will remedy itself.

This also reduces the amount of failures to connect to Andromeda service
as one static variable has been removed.

Signed-off-by: Nicholas Chum <nicholas@prjkt.io>