FileOperations: Consume exception when trying to copy non-existent files
When trying to use the AssetManager to open a file that doesn't exist,
for example the 'version' file in individual overlay directories,
an unhandled FileNotFoundException is thrown which does not break
execution like it should, rather letting other actual copy calls to
throw an IOException and return. This causes unnecessary logspam as
well wastes precious execution time. Bail as soon as we verify that
the file doesn't exist and continue with other things.

Signed-off-by: Harsh Shandilya <harsh@prjkt.io>