References: Don't crash on unrooted devices
Unlike Kotlin's lazy properties, all members of
a Java class initialize immediately when the application
is started which causes the app to look for root to
try and establish a Magisk directory which fails on
unrooted devices and causes an exception. This fix is
far from ideal but will do for now.

Signed-off-by: Harsh Shandilya <harsh@prjkt.io>